pyproject.toml
[project]
name = "frejun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.19",
    "djangorestframework>=3.15.2",
    "factory-boy>=3.3.3",
    "faker>=36.2.2",
    "python-dotenv>=1.0.1",
]
